requirements/requirements_dev.txt
certifi==2017.11.5
chardet==3.0.4
codecov==2.0.10
configparser==3.5.0
coverage==4.4.2
Django==1.11.7
django-filter==1.0.4
djangorestframework==3.7.3
enum34==1.1.6
flake8==3.4.1
flake8-isort==2.2.1
flake8-polyfill==1.0.1
idna==2.6
isort==4.2.15
mccabe==0.6.1
model-mommy==1.4.0
pycodestyle==2.3.1
pyflakes==1.5.0
pytz==2017.3
requests==2.18.4
six==1.11.0
testfixtures==5.1.1
urllib3==1.22
